+++ LuckParser/ConsoleProgram.cs
@@ -39,10 +39,6 @@ namespace LuckParser
 
         private void ParseLog(object logFile)
         {
-            UploadController up_controller = new UploadController();
-            System.Globalization.CultureInfo before = Thread.CurrentThread.CurrentCulture;
-            Thread.CurrentThread.CurrentCulture =
-                    new System.Globalization.CultureInfo("en-US");
             GridRow row = new GridRow(logFile as string, "Ready to parse")
             {
                 BgWorker = new System.ComponentModel.BackgroundWorker()
@@ -52,54 +48,7 @@ namespace LuckParser
             };
             row.Metadata.FromConsole = true;
 
-            FileInfo fInfo = new FileInfo(row.Location);
-            if (!fInfo.Exists)
-            {
-                throw new CancellationException(row, new FileNotFoundException("File does not exist", fInfo.FullName));
-            }
-            try
-            {
-                ParsingController control = new ParsingController();
-
-                if (!GeneralHelper.HasFormat())
-                {
-                    throw new CancellationException(row, new InvalidDataException("No output format has been selected"));
-                }
-
-                if (GeneralHelper.IsSupportedFormat(fInfo.Name))
-                {
-                    //Process evtc here
-                    ParsedLog log = control.ParseLog(row, fInfo.FullName);
-                    string[] uploadresult = up_controller.UploadOperation(row, fInfo);
-                    //Creating File
-                    GeneralHelper.GenerateFiles(log, row, uploadresult, fInfo);
-                }
-                else
-                {
-                    row.BgWorker.UpdateProgress(row, "Not EVTC", 100);
-                    Console.Error.Write("Not EVTC");
-                    throw new CancellationException(row, new InvalidDataException("Not EVTC"));
-                }
-            }
-            catch (SkipException s)
-            {
-                Console.Error.Write(s.Message);
-                throw new CancellationException(row, s);
-            }
-            catch (TooShortException t)
-            {
-                Console.Error.Write(t.Message);
-                throw new CancellationException(row, t);
-            }
-            catch (Exception ex) when (!System.Diagnostics.Debugger.IsAttached)
-            {
-                Console.Error.Write(ex.Message);
-                throw new CancellationException(row, ex);
-            } 
-            finally
-            {
-                Thread.CurrentThread.CurrentCulture = before;
-            }
+            ProgramHelper.DoWork(row);
             
         }
     }
